Moved membership stuff to new app; made migrations
