+++ members/apps.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class MembersConfig(AppConfig):
+    name = 'members'
