+++ members/forms.py
@@ -1,0 +1,35 @@
+from django import forms
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.models import User
+
+from members.models import Person
+
+
+class PersonForm(forms.ModelForm):
+    def __init__(self, *args, **kwargs):
+        super(PersonForm, self).__init__(*args, **kwargs)
+        self.fields['email'].required = True
+        self.fields['lab'].required = True
+        self.fields['lab'].label = 'Affiliation/Lab'
+    class Meta:
+        model=Person
+        fields = ('first_name','last_name','email','lab','year')
+
+class RegistrationForm(UserCreationForm):
+    email = forms.EmailField(required=True)
+    first_name = forms.CharField(max_length=30,required=True)
+    last_name = forms.CharField(max_length=30,required=True)
+
+    class Meta:
+        model = User
+        fields = ('username', 'email', 'first_name', 'last_name', 'password1', 'password2')
+
+    def save(self, commit=True):
+        user = super(RegistrationForm, self).save(commit=False)
+        user.email = self.cleaned_data['email']
+        user.first_name = self.cleaned_data['first_name']
+        user.last_name = self.cleaned_data['last_name']
+
+        if commit:
+            user.save()
+        return user
