+++ website/admin.py
@@ -2,17 +2,12 @@
 from __future__ import unicode_literals
 
 from django.contrib import admin
-from django.contrib.auth.admin import UserAdmin
-from django.contrib.auth.forms import UserCreationForm, UserChangeForm
-from django.contrib.auth.models import User
 from django.shortcuts import redirect
-from django.utils.safestring import mark_safe
 from django.urls import reverse
 from django.conf import settings
 
-# Register your models here.
-from .models import *
-from .forms import *
+from website.forms import PresentationForm
+from website.models import *
 
 class ExtraMedia:
     # Use in form if needed
@@ -149,99 +144,11 @@ class PresentationAdmin(redirectFromAdmin):
             form.base_fields['event'].help_text = 'No session specified in URL, assuming active session. If not, you can set via Session -> Event -> details'
         return form
 
-def make_subscribed(modeladmin, request, queryset):
-    queryset.update(subscribed=True)
-    return redirect(reverse('admin:website_person_changelist'))
-make_subscribed.short_description = 'Mark selected people subscribed'
-
-class PersonAdmin(redirectFromAdmin):
-    readonly_fields = ('join_method',)
-    list_display = ('__unicode__', 'subscribed','lab','email','year','join_method')
-    list_filter = ('subscribed','year','join_method',)
-    actions = (make_subscribed,)
-
-# Update User admin to include profile inline
-class ProfileInline(admin.StackedInline):
-    model = Profile
-    can_delete = False
-    verbose_name_plural = 'Profile'
-    fk_name = 'user'
-
-class EmailRequiredMixin(object):
-    def __init__(self, *args, **kwargs):
-        super(EmailRequiredMixin, self).__init__(*args, **kwargs)
-        # make user email field required
-        if 'first_name' in self.fields:
-            self.fields['first_name'].required = True
-        if 'last_name' in self.fields:
-            self.fields['last_name'].required = True
-        if 'email' in self.fields:
-            self.fields['email'].required = True
-class MyUserCreationForm(UserCreationForm):
-    pass
-class MyUserChangeForm(EmailRequiredMixin, UserChangeForm):
-    pass
-class CustomUserAdmin(UserAdmin):
-    form = MyUserChangeForm
-    add_form = MyUserCreationForm
-    staff_fieldsets = (
-        (None, {'fields': ('username', 'password')}),
-        (('Personal info'), {'fields': ('first_name', 'last_name', 'email')}),
-    )
-    list_filter = () # Small group, not necessary
-    inlines = (ProfileInline, )
-    list_display = ('username','get_role', 'email', 'first_name', 'last_name', 'is_staff')
-    list_select_related = ('profile', ) # Streamline database queries
-
-    def get_role(self,instance):
-        return instance.profile.role
-    get_role.short_description = 'Role'
-
-    def get_inline_instances(self, request, obj=None):
-        if not obj:
-            return list()
-        return super(CustomUserAdmin, self).get_inline_instances(request, obj)
-
-    def change_view(self, request, id, *args, **kwargs):
-        # for non-superuser [NOTE this does not provide security, just niceer view]
-        if not request.user.is_superuser:
-            try:
-                self.fieldsets = (None, {'fields': ()}),
-                try:
-                    if request.user == User.objects.get(id=id):
-                        self.fieldsets = self.staff_fieldsets
-                except User.DoesNotExist:
-                    pass
-                response = super(CustomUserAdmin, self).change_view(request, id, *args, **kwargs)
-            finally:
-                # Reset fieldsets to its original value
-                self.fieldsets = UserAdmin.fieldsets
-            return response
-        else:
-            return super(CustomUserAdmin, self).change_view(request, id, *args, **kwargs)
-
-    def get_queryset(self, request):
-        qs = super(UserAdmin, self).get_queryset(request)
-        if not request.user.is_superuser:
-            return qs.filter(is_superuser=False)
-        return qs
-
-    def save_model(self, request, obj, form, change):
-        obj.is_staff = True
-        obj.save()
-
 admin.site.register(Donor,DonorAdmin)
 admin.site.register(IQUISE,IQUISEAdmin)
 admin.site.register(Session,SessionAdmin)
 admin.site.register(Event,EventAdmin)
 admin.site.register(Presenter,PresenterAdmin)
 admin.site.register(Presentation,PresentationAdmin)
-admin.site.register(Person,PersonAdmin)
-admin.site.register(Department)
-admin.site.register(School)
 admin.site.register(EmbeddedVideo,EmbeddedVideoAdmin)
 admin.site.register(EmbedEngine)
-
-# Reset admin User
-admin.site.unregister(User)
-admin.site.register(User, CustomUserAdmin)
