+++ website/urls.py
@@ -6,10 +6,6 @@ app_name = 'website'
 urlpatterns = [
     url(r'^$', views.index, name='index'),
     url(r'^presentation/(?P<presentation_id>[0-9]+)/$', views.presentation, name='presentation'),
-    url(r'^leadership/$', views.people, name='people_deprecated'), # deprecated, but don't delete so existing hyperlinks stil work
-    url(r'^exec/$', views.people, name='people'),
-    url(r'^join/$',views.join.as_view(), name='join'),
     url(r'^archive/$',views.archive, name='archive'),
-    url(r'^register/(?P<hash>.*)/$', views.staff_register, name="register"),
     url(r'^scheduler/$', views.scheduler, name='scheduler'),
 ]
